Tests/LibWeb/Text/expected/DOM/Elements-from-point.txt

  Negative coordinates return empty array: true
Coordinates outside the viewport return empty array: true
== Elements at (500, 10) ==
<DIV id="large-box" >
<HTML >
== FIXME: Elements at (550, 60) ==
<DIV id="small-box" >
<DIV id="large-box" >
<DIV id="large-box" >
<DIV id="large-box" >
<DIV id="small-box" >
<PRE id="out" >
<HTML >
